# examples/path_tracing.v: faster floating point comparisons