--- a/.gitignore
+++ b/.gitignore
@@ -57,3 +57,4 @@ info.log
 # vim/emacs editor backup files
 *~
 thirdparty/freetype/
+*.ppm