1. This PR parametrizes examples/path_tracing.v, thus making it
more suitable for comparisons with the smallpt original.

2. Implements support for -d no_bounds_checking
    `v -prod examples/path_tracing.v`
    ... and then:
    `time ./examples/path_tracing`
    => ~2.083s on average
        
    With `v -d no_bounds_checking -prod examples/path_tracing.v`
    ... then again:
    `time ./examples/path_tracing`
    => ~1.995s on average
    
    ===> TLDR: on average, NOT doing bounds checking,
    speeds up examples/path_tracing.v by ~88ms (~4.4%) .

3. Uses macro_fXX_YY C macros for f32 and f64 comparisons.

   Before: `./v -prod examples/path_tracing.v`, then `time
   ./examples/path_tracing` on average takes 2.620s .

   After:  `./v -prod examples/path_tracing.v`, then `time
   ./examples/path_tracing` on average takes 2.084s .

   ===> TLDR: the average speedup from this change for 
   path_tracing.v is ~20% .

4. Cleans up formatting and tweaks constants in 
examples/path_tracing.v so that it is more close to the original.

5. TODO: make the output of examples/path_tracing.v identical to 
the one produced by smallpt.

path_tracing produces:
[![0-64-image.png](https://i.postimg.cc/fysKW3vZ/0-64-image.png)](https://postimg.cc/njd70h5w)

... while smallpt produces:
![smallpt 64](https://url4e.com/gyazo/images/634b865d.png)


Usage:
```shell
0[00:09:51]delian@nemesis: /v/nv $ time examples/path_tracing 64
0.64.image.ppm 0 1024 768
Rendering (64 spp) 100.00%
Rendering finished. Took: 84267ms
Image saved as [0.64.image.ppm]. Took:   465ms

real    1m24.742s
user    1m23.875s
sys     0m0.082s
0[00:11:19]delian@nemesis: /v/nv $
```

*Without OpenMP*, now the v port is at ~ the same speed as smallpt.
I.e. running
```shell
g++ -flto -O3 -ffast-math -funroll-all-loops -march=native smallpt.cpp -o smallpt
time smallpt 64
```
... shows that smallpt takes ~1m36.090s .

Using OpenMP of course still makes smallpt much faster than the V
version (50s for 2 cores, vs 1m24.742s for 1 core for V).