@spytheman 2020-02-15
Another example, with scene 2 this time:
`time examples/path_tracing 32 scene_2.32_samples.image.ppm 2 1024 768`