@spytheman 2020-02-15
And scene 1:
`time examples/path_tracing 32 scene_1.32_samples.image.ppm 1 1024 768`
=> `Rendering finished. Took: 27026ms`

[![scene-1-32-samples-image.png](https://i.postimg.cc/MH41Mv7v/scene-1-32-samples-image.png)](https://postimg.cc/n94XNVwJ)