@penguindark 2020-02-15
@spytheman Have you tested using the full precision instead of the tabelized sin/cos?