@spytheman 2020-02-15
@penguindark Yes, it produces the same kind of picture, but is ~30% slower on my PC.