@spytheman 2020-02-15
I am puzzled why is that - the algo seems 1:1 the same as the one from smallpt, even the constants are now equal...