@penguindark 2020-02-15
probably some sign or approximation, I'll check as I've some spare time :)