@spytheman 2020-02-16
Current output for: 
`time examples/path_tracing 512 a.512.image.ppm 0 1024 768`
[![a-512-image.png](https://i.postimg.cc/WtRgw880/a-512-image.png)](https://postimg.cc/Pv4CtmBq)