@spytheman 2020-02-16
p.s. after the fix, the V version slowed down slightly, and now there is virtually no performance difference between smallpt and the v version in single threaded mode.
1m34.350s for V 
1m36.090s for the C version