@penguindark 2020-02-16
@spytheman by the way, we must think about f32/f64 auto conversion.
This test on this simple code has show us the problem with the conversion and what can happen.
I think we should involve @medvednikov in this.
*P.S. I write it for this scope ;)*