@medvednikov 2020-02-16
V is also using f32 by default right now, but it will be f64 in the future.

I regret not fixing this as early as possible.