@spytheman 2020-02-16
(rebased over current master)